refactor: apis 기능 분리 진행
