+++ server/apis/swagger.py
@@ -1,0 +1,20 @@
+import secrets
+from fastapi import HTTPException, Depends, status
+from fastapi.security import HTTPBasic, HTTPBasicCredentials
+from core.config import settings
+
+
+# HTTP 기본 인증을 사용하는 Security 객체 생성
+security = HTTPBasic()
+
+# Swagger 인증
+def get_current_username(credentials: HTTPBasicCredentials = Depends(security)) -> str:
+    correct_username = secrets.compare_digest(credentials.username, settings.ADMIN_USERNAME)
+    correct_password = secrets.compare_digest(credentials.password, settings.ADMIN_PASSWORD)
+    if not (correct_username and correct_password):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect email or password",
+            headers={"WWW-Authenticate": "Basic"},
+        )
+    return credentials.username
